requirements.txt
bcrypt==4.2.1
celery==5.4.0
Django==5.1.4
django-environ==0.11.2
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
psycopg2-binary==2.9.10
redis==5.2.1